+++ libfhqpm/utils_shell.py
@@ -75,14 +75,14 @@ class UtilsShell:
                     _output.write("ERROR: returncode " + str(_returncode) + "\n")
                 sys.exit(_returncode)
             return
-        sys.exit("Could not start process")
+        print("ERROR: Could not start process")
+        sys.exit(-1)
 
     @staticmethod
     def run_command_get_output(_log, _command):
         """ run_command_get_output """
         # _log.info("Run command: " + " ".join(_command))
         _output = ""
-        _returncode = -1
         with subprocess.Popen(
             _command,
             stdout=subprocess.PIPE,