+++ src/database/fhq_server_database_file.h
@@ -32,8 +32,8 @@
 #pragma once
 
 #include <cstring>
-#include <string>
 #include <mutex>
+#include <string>
 #include <vector>
 
 class FhqServerDatabaseFileUpdateInfo {