+++ CHANGELOG.md
@@ -5,7 +5,29 @@ All notable changes to fhq-server project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [v0.2.53] - 2025-08-10 (2024 Jule 21)
+## [v0.2.54] - 2025-08-21 (2025 Aug 21)
+
+- Fix Dockerfile
+- [web-admin] Fix login dialog
+- Fix game_update and game_update_logo
+- Removed ci/travis/run2_language_node_js.sh
+- Fix warning: return in src/database/fhq_server_database_file.cpp
+- Fix clang-format
+- Added new task for runtime migration data (update missing uuids)
+- Removed unsed and deprecated options
+- Fix TODO deprecated setWorkDir and fix behaivor when not set '-wd'
+- Fix pylint warnings
+- Fixed 'Unreachable code' in 'tests/server-api-tests/test_01_games.py' by code-scanning
+- Fixed 'Variable defined multiple times' in 'libfhqpm/command_code_stats.py' by code-scanning
+- Fixed 'Variable defined multiple times ' in 'libfhqpm/utils_shell.py' by code-scanning
+- Fixed 'Unreachable code' in 'libfhqpm/utils_shell.py' by code-scanning
+- Fix warning
+- Updated .vscode/settings.json
+- Added new field uuid to public_event table
+- Added EmployPublicEvents and redesing work with this employ in cmd
+- Added model_public_event
+
+## [v0.2.53] - 2025-08-10 (2025 Jule 21)
 
 - Updated wsjcpp-yaml from 0.1.7 to 0.1.8
 - Updated years to -2025